Improved DeviceStreamService connection handling
