+++ frontend/patient/src/services/DeviceStreamService.js
@@ -1,5 +1,6 @@
 import { Stomp } from '@stomp/stompjs/esm6/compatibility/stomp';
 import { WS_URL } from '../config';
+import authService from './AuthService';
 
 const url = `${WS_URL}/requests`;
 
@@ -8,9 +9,24 @@ webSocket.debug = () => {};
 webSocket.reconnect_delay = 1000;
 
 let token = null;
+let connected = false;
 
-const subscribe = (newToken, setForm, cancelForm) => {
-  token = newToken;
+const subscribe = async (history) => {
+  if (connected) return;
+
+  token = localStorage.getItem('device-token');
+  if (!token) { return; }
+  try { await authService.meDevice(token); } catch (e) { localStorage.removeItem('device-token'); history.push('/'); return; }
+
+  const setForm = (pinCode) => {
+    authService.login(pinCode)
+      .then((c) => localStorage.setItem('credentials', JSON.stringify(c)))
+      .then(() => history.push('/form'));
+  };
+  const cancelForm = () => {
+    localStorage.removeItem('credentials');
+    history.push('/');
+  };
 
   const handleRequest = (message) => {
     const request = JSON.parse(message.body);
@@ -27,6 +43,7 @@ const subscribe = (newToken, setForm, cancelForm) => {
 
   webSocket.connect({ Authorization: `Bearer ${token}` }, () => {
     webSocket.subscribe('/forms/', handleRequest);
+    connected = true;
   });
 };
